requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
Flask-Migrate==3.1.0
psycopg2-binary==2.9.2
Flask-JWT-Extended==4.3.1
cloudinary==1.25.0
bcrypt==3.2.0
python-dotenv==0.19.2
Flask-Cors==5.0.1
gunicorn==20.1.0
Werkzeug==2.2.3
Flask-Mail==0.9.1
requests==2.28.1
Pillow==9.3.0
